src/requirements.txt
celery==5.0.5
Django==3.1.6
django-cors-headers==3.7.0
django-environ==0.4.5
django-extensions==3.1.1
# No imperative filtering, declarative beautiful filters
django-filter==2.4.0
# For process specification
django-fsm==2.7.1
django-jazzmin==2.4.4
djangorestframework==3.12.2
django-reversion==3.0.9
django-silk==4.1.0
django-solo==1.1.5
gunicorn==20.0.4
gevent==21.1.2
ipython==7.20.0
packaging==20.9
psycopg2-binary==2.8.6
requests==2.25.1
sentry-sdk==0.20.1
certifi==2020.12.5
chardet==4.0.0
cx-Oracle==8.1.0
idna==2.10
numpy==1.20.2
pandas==1.2.3
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
urllib3==1.26.4
django-graphiql==0.4.4
graphene==2.1.8
graphene-django==2.15.0